Update cachematrix.R
These two functions allow the user to perform repetitive matrix inverse operations faster by using a cache to store the inverse. Subsequent calls check to see if the inverse is in the cache (i.e., has already been calculated). If so, it just returns the stored inverse instead of recalculating it.